requirements.txt
pyinstaller>=6.11.1
PySide6_Essentials>=6.8.1
shiboken6>=6.8.1
